requirements.txt
discord.py>=1.5.1
requests>=2.25.1
discord-py-slash-command>=1.1.2
youtube_dl>=2021.4.1
